fix final transitivity test case
To do so, I added the original function's name as an attribute to the decorator wrapped function so that that can be used as a key to the global table instead. This name will be accessible to new references to the object as well, so the actual function name doesn't matter anymore.